administradorseries/src/app/serie/serie-list/serie-list.component.ts
import { Component, OnInit } from '@angular/core';
import { Serie } from '../serie';
import { SerieService } from '../serie.service';

@Component({
  selector: 'app-serie-list',
  templateUrl: './serie-list.component.html',
  styleUrls: ['./serie-list.component.css']
})
export class SerieListComponent implements OnInit {

  series: Array<Serie> = [];
  message: string = "";
  constructor(private serieService: SerieService) { }

  getSeries(): void {
    this.serieService.getSeries().subscribe((series) => {
      this.series = series;
      this.message = this.getAverageSeasons(series);
    });
  }

  getAverageSeasons(series: Serie[]): string {
    let totalSeasons: number = 0;
    series.forEach((serie) => totalSeasons = totalSeasons + serie.seasons);
    return `Average number of seasons: ${totalSeasons / series.length}`;
  }

  ngOnInit(): void {
    this.getSeries();
  }

}
